Move actual detection into library
